error removed

+++ energy_efficiency/pipeline/pipeline.py
@@ -200,7 +200,7 @@ class Pipeline(Thread):
         try:
             if os.path.exists(Pipeline.experiment_file_path):
                 df = pd.read_csv(Pipeline.experiment_file_path)
-                limit = -1 * float(limit)
+                limit = -1 * int(limit)
                 return df[limit:].drop(columns=["experiment_file_path", "initialization_timestamp"], axis=1)
             else:
                 return pd.DataFrame()
